requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Cors
pandas
numpy
scikit-learn
requests
SQLAlchemy
gunicorn
psycopg2-binary
python-dotenv